# Speedup .SequenceCompareTo(byte, ...)