@benaadams 2019-01-22
Test extensions https://github.com/dotnet/corefx/pull/34742