@benaadams 2019-01-24
Added performance numbers and rebased