@redknightlois 2019-01-24
@benaadams can you run with both with cache misses and without cache misses? Also with high numbers like 32K to 1M (with a few in-between)