@benaadams 2019-01-24
> Also with high numbers like 32K to 1M (with a few in-between)

Different

```
|  Method |       N |      Mean |      Error |    StdDev |
|-------- |-------- |----------:|-----------:|----------:|
|  Native |    1024 | 11.928 ns |  4.9578 ns | 0.2718 ns |
| Managed |    1024 |  4.620 ns |  3.3234 ns | 0.1822 ns |

|  Native |   32768 | 10.927 ns | 11.8992 ns | 0.6522 ns |
| Managed |   32768 |  7.585 ns | 31.4300 ns | 1.7228 ns |

|  Native |   65536 | 12.784 ns |  5.7687 ns | 0.3162 ns |
| Managed |   65536 |  4.524 ns |  2.8635 ns | 0.1570 ns |

|  Native |  131072 | 10.691 ns |  3.8721 ns | 0.2122 ns |
| Managed |  131072 |  4.438 ns |  0.4294 ns | 0.0235 ns |

|  Native |  262144 | 10.326 ns |  0.5011 ns | 0.0275 ns |
| Managed |  262144 |  4.971 ns |  0.3389 ns | 0.0186 ns |

|  Native |  524288 | 10.368 ns |  0.3761 ns | 0.0206 ns |
| Managed |  524288 |  4.477 ns |  0.1690 ns | 0.0093 ns |

|  Native | 1048576 | 10.389 ns |  1.4737 ns | 0.0808 ns |
| Managed | 1048576 |  5.013 ns |  0.4726 ns | 0.0259 ns |

|  Native | 8388608 | 10.503 ns |  3.6709 ns | 0.2012 ns |
| Managed | 8388608 |  4.545 ns |  1.6608 ns | 0.0910 ns |
```

AlmostSameButDifferent

```
|  Method |       N |            Mean |            Error |         StdDev |
|-------- |-------- |----------------:|-----------------:|---------------:|
|  Native |    1024 |        52.23 ns |         6.140 ns |      0.3366 ns |
| Managed |    1024 |        33.61 ns |        16.190 ns |      0.8875 ns |

|  Native |   32768 |     2,008.78 ns |        67.170 ns |      3.6818 ns |
| Managed |   32768 |     1,114.52 ns |        83.492 ns |      4.5765 ns |

|  Native |   65536 |     4,011.11 ns |       598.936 ns |     32.8297 ns |
| Managed |   65536 |     2,251.27 ns |       626.958 ns |     34.3657 ns |

|  Native |  131072 |     8,021.91 ns |     1,162.816 ns |     63.7379 ns |
| Managed |  131072 |     5,016.38 ns |       122.943 ns |      6.7389 ns |

|  Native |  262144 |    16,741.14 ns |       737.067 ns |     40.4011 ns |
| Managed |  262144 |    12,111.76 ns |     1,401.922 ns |     76.8441 ns |

|  Native |  524288 |    33,311.27 ns |     3,603.438 ns |    197.5166 ns |
| Managed |  524288 |    24,386.69 ns |     5,319.465 ns |    291.5778 ns |

|  Native | 1048576 |    66,513.58 ns |     7,202.769 ns |    394.8081 ns |
| Managed | 1048576 |    50,523.14 ns |    13,804.180 ns |    756.6537 ns |

|  Native | 8388608 | 1,779,989.61 ns |   805,216.188 ns | 44,136.6164 ns |
| Managed | 8388608 | 1,806,247.14 ns | 1,316,669.259 ns | 72,171.0850 ns |
```
Same

```
|  Method |       N |            Mean |           Error |         StdDev |
|-------- |-------- |----------------:|----------------:|---------------:|
|  Native |    1024 |        50.10 ns |       3.5770 ns |      0.1961 ns |
| Managed |    1024 |        22.84 ns |       0.1028 ns |      0.0056 ns |

|  Native |   32768 |     1,999.43 ns |     356.4009 ns |     19.5355 ns |
| Managed |   32768 |     1,098.49 ns |      32.1779 ns |      1.7638 ns |

|  Native |   65536 |     4,012.88 ns |     431.7510 ns |     23.6657 ns |
| Managed |   65536 |     2,356.23 ns |   4,815.9938 ns |    263.9809 ns |

|  Native |  131072 |     7,987.18 ns |     342.8520 ns |     18.7929 ns |
| Managed |  131072 |     4,996.34 ns |     211.1606 ns |     11.5744 ns |

|  Native |  262144 |    16,504.62 ns |   1,353.8718 ns |     74.2103 ns |
| Managed |  262144 |    11,966.56 ns |   1,042.4223 ns |     57.1387 ns |

|  Native |  524288 |    32,981.75 ns |   2,399.1959 ns |    131.5080 ns |
| Managed |  524288 |    23,830.30 ns |   2,754.7899 ns |    150.9993 ns |

|  Native | 1048576 |    66,725.47 ns |   7,043.4319 ns |    386.0743 ns |
| Managed | 1048576 |    48,112.26 ns |   7,447.5663 ns |    408.2262 ns |

|  Native | 8388608 | 1,667,792.22 ns |  49,976.4103 ns |  2,739.3757 ns |
| Managed | 8388608 | 1,613,212.61 ns | 252,499.1025 ns | 13,840.3278 ns |
```

> can you run with both with cache misses and without cache misses?

If you can come up with a benchmark for that where killing the cache doesn't dominate. However, it should effect both similarly?