@benaadams 2019-01-24
@dotnet-bot test OSX10.12 x64 Checked Innerloop Build and Test please